controller/requirements.txt
adafruit-circuitpython-neopixel==6.3.11
rpi-ws281x==5.0.0
